Add c++ test files for target tests
All files residing in a directory 'test' should be considered tests and
shall also be built automatically, as long as there is a CMakeLists.txt
file saying so, because tests also needs special compile and link flags.

The purpose with a naming convention is to use a macro that just builds
all files into separate executables that can later be executed as part
of test running.